requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
discord.py==1.6.0
idna
markovify==0.9.0
multidict==5.1.0
nekos.py==1.0.3
Pillow==8.1.2
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
typing-extensions==3.7.4.3
Unidecode==1.2.0
urllib3==1.26.3
websockets==8.1
yarl==1.6.3
